requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.2.5
django-formtools==2.5.1
django-phonenumber-field==8.1.0
django-ranged-response==0.2.0
django-recaptcha==4.1.0
django-tinymce==4.1.0
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
jsmin==3.0.1
packaging==25.0
pillow==11.0.0
psycopg2-binary==2.9.10
pypng==0.20220715.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
stripe==11.2.0
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.9.0
